removed files from META-INF can be configured in the extension
INDEX.LIST and SIG-* files are also removed from copied files. The
Jnlp-Classloader does have issues with jar files having a index.list file
if the file name does not match the name mentioned in the index.list. An
index.list in a jar other than the main jar file does not make sense,
therefore I think it is best to remove this for now.

In a later commit I want to allow generating a index.list in the main jar
file.