feat: add config option to clean all named sections from manifest
Add new config option `signJarRemovedNamedManifestEntries`. This option
is used for the named manifest attributes, while the old
`signJarRemovedManifestEntries` is used for the mainAttributes only.

When after removing manifest entries a named entry has no attributes, the
entry is removed as well.

Closes #34

BREAKING CHANGE:
Config option `signJarRemovedManifestEntries` is not used for the named
attributes any more. For these entries the new
`signJarRemovedNamedManifestEntries` is used.