feat: Respect $JAVA_HOME when calling pack200
When pack200 is not inside the $PATH, the task failed. Now if the System
environment variable JAVA_HOME is set, it is used to fully qualify the
path to pack200 and unpack200 ($JAVA_HOME/bin/pack200)

If $JAVA_HOME is not set, `System.getProperty('java.home') is used. Since
`java.home` points to the jre subdirectory within the JDK, this part
of `java.home` is removed in that case.

Closes #33