feat: allow ignoring duplicate entries in Jar files
Add new member `duplicatesStrategy` to task signJars.

This member is unset by default, which is the same behavior as before.
Setting the value to `INCLUDE` would result in the same.

If the member is set to `EXCLUDE`, duplicate entries in jar files are
ignored and a debug message is printed to the log.

If it is `WARN`, a warning is printed to the log, but the build will fail
nonetheless, because the JarOutputStream will not allow duplicate entries.

If it is `FAIL`, the build will fail as well, but gracefully by throwing a
DuplicateFileCopyingException.

Closes: #37