fix(SignJars): works with Java 8u141
The gradle-jnlp-plugin removes all named entries `.*-Digest` and some main
attributes of the MANIFEST.MF to prevent some type of errors.

In doing this, it does alter the manifest of the original jar file to
create a new jar file. But it seems, that 8u141 does has an issue with
this.

The SignJarsTask now creates a new Manifest instead of copying the original
one.

Closes #42